:card_file_box: update table structure
made sqlite tables have a simpler structure reducing columns and better validation for duplicate entries.

also add new attribute class specifically for sqlcolumn name conversions. so it's clear that the attribute is used for reflection(?) of the model's (class') sql table and column.
since each model represents a table and it's properties the columns